modules/sc-mesh-secure-deployment/src/authentication/requirements.txt
paramiko==2.7.2
wifi==0.3.8
pandas==1.1.5
netifaces==0.10.9
numpy==1.19.5
ruamel.yaml
